// src/lib/supabase/database.types.ts
export type Database = {
	// Allows to automatically instantiate createClient with right options
	// instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
	__InternalSupabase: {
		PostgrestVersion: '13.0.5';
	};
	public: {
		Tables: {
			games: {
				Row: {
					black_player_id: string | null;
					created_at: string | null;
					id: string;
					result: Database['public']['Enums']['game_result'] | null;
					round_id: string | null;
					white_player_id: string | null;
				};
				Insert: {
					black_player_id?: string | null;
					created_at?: string | null;
					id?: string;
					result?: Database['public']['Enums']['game_result'] | null;
					round_id?: string | null;
					white_player_id?: string | null;
				};
				Update: {
					black_player_id?: string | null;
					created_at?: string | null;
					id?: string;
					result?: Database['public']['Enums']['game_result'] | null;
					round_id?: string | null;
					white_player_id?: string | null;
				};
				Relationships: [
					{
						foreignKeyName: 'games_black_player_id_fkey';
						columns: ['black_player_id'];
						isOneToOne: false;
						referencedRelation: 'leaderboard';
						referencedColumns: ['id'];
					},
					{
						foreignKeyName: 'games_black_player_id_fkey';
						columns: ['black_player_id'];
						isOneToOne: false;
						referencedRelation: 'players';
						referencedColumns: ['id'];
					},
					{
						foreignKeyName: 'games_round_id_fkey';
						columns: ['round_id'];
						isOneToOne: false;
						referencedRelation: 'rounds';
						referencedColumns: ['id'];
					},
					{
						foreignKeyName: 'games_white_player_id_fkey';
						columns: ['white_player_id'];
						isOneToOne: false;
						referencedRelation: 'leaderboard';
						referencedColumns: ['id'];
					},
					{
						foreignKeyName: 'games_white_player_id_fkey';
						columns: ['white_player_id'];
						isOneToOne: false;
						referencedRelation: 'players';
						referencedColumns: ['id'];
					}
				];
			};
			players: {
				Row: {
					created_at: string | null;
					id: string;
					is_active: boolean | null;
					name: string;
					rating: number | null;
				};
				Insert: {
					created_at?: string | null;
					id?: string;
					is_active?: boolean | null;
					name: string;
					rating?: number | null;
				};
				Update: {
					created_at?: string | null;
					id?: string;
					is_active?: boolean | null;
					name?: string;
					rating?: number | null;
				};
				Relationships: [];
			};
			rounds: {
				Row: {
					created_at: string | null;
					id: string;
					is_current: boolean | null;
					round_number: number;
					status: string | null;
				};
				Insert: {
					created_at?: string | null;
					id?: string;
					is_current?: boolean | null;
					round_number: number;
					status?: string | null;
				};
				Update: {
					created_at?: string | null;
					id?: string;
					is_current?: boolean | null;
					round_number?: number;
					status?: string | null;
				};
				Relationships: [];
			};
			leaderboard: {
				Row: {
					buchholz: number | null;
          is_active: boolean | null;
					id: string | null;
					name: string | null;
					score: number | null;
					wins: number | null;
          losses: number | null;
          draws: number | null;
				};
				Insert: {
					buchholz: number | null;
          is_active: boolean | null;
					id: string | null;
					name: string | null;
					score: number | null;
					wins: number | null;
          losses: number | null;
          draws: number | null;
				};
				Update: {
					buchholz: number | null;
          is_active: boolean | null;
					id: string | null;
					name: string | null;
					score: number | null;
					wins: number | null;
          losses: number | null;
          draws: number | null;
				};
				Relationships: [];
			};
		};
		Views: {
			leaderboard: {
				Row: {
          id: string | null;
					name: string | null;
					score: number | null;
					wins: number | null;
          losses: number | null;
          draws: number | null;
					buchholz: number | null;
          is_active: boolean | null;
				};
				Relationships: [];
			};
			player_game_points: {
				Row: {
					opponent_id: string | null;
					player_id: string | null;
					points: number | null;
				};
				Relationships: [];
			};
		};
		Functions: {
			[_ in never]: never;
		};
		Enums: {
			game_result: 'white_wins' | 'black_wins' | 'draw' | 'bye' | 'scheduled';
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>];
// ...
export type Enums<
	DefaultSchemaEnumNameOrOptions extends
		| keyof DefaultSchema['Enums']
		| { schema: keyof DatabaseWithoutInternals },
	EnumName extends DefaultSchemaEnumNameOrOptions extends {
		schema: keyof DatabaseWithoutInternals;
	}
		? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
		: never = never
> = DefaultSchemaEnumNameOrOptions extends {
	schema: keyof DatabaseWithoutInternals;
}
	? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
	: DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
	? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
	: never;

export type CompositeTypes<
	PublicCompositeTypeNameOrOptions extends
		| keyof DefaultSchema['CompositeTypes']
		| { schema: keyof DatabaseWithoutInternals },
	CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
		schema: keyof DatabaseWithoutInternals;
	}
		? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes']
		: never = never
> = PublicCompositeTypeNameOrOptions extends {
	schema: keyof DatabaseWithoutInternals;
}
	? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes'][CompositeTypeName]
	: PublicCompositeTypeNameOrOptions extends keyof DefaultSchema['CompositeTypes']
	? DefaultSchema['CompositeTypes'][PublicCompositeTypeNameOrOptions]
	: never;